requirements.txt
google-generativeai>=0.8.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.140.0
requests>=2.32.0
pydantic>=2.8.0
python-dotenv>=1.0.0
Pillow>=10.4.0
